+++ kibot/kiplot.py
@@ -155,10 +155,15 @@ def _run_command(command, change_to):
     return run(command, check=True, stdout=PIPE, stderr=STDOUT, cwd=change_to)
 
 
-def run_command(command, change_to=None, just_raise=False, use_x11=False, err_msg=None, err_lvl=FAILED_EXECUTE):
+def run_command(command, change_to=None, just_raise=False, use_x11=False, err_msg=None, err_lvl=FAILED_EXECUTE,
+                force_en=False):
     logger.debug('- Executing: '+GS.pasteable_cmd(command))
     if change_to is not None:
         logger.debug('- CWD: '+change_to)
+    old_lang = None
+    if force_en:
+        old_lang = os.environ['LANG']
+        os.environ['LANG'] = 'en'
     try:
         if use_x11 and not GS.on_windows:
             logger.debug('Using Xvfb to run the command')
@@ -173,6 +178,10 @@ def run_command(command, change_to=None, just_raise=False, use_x11=False, err_ms
         if err_msg is not None:
             err_msg = err_msg.format(ret=e.returncode)
         GS.exit_with_error(err_msg, err_lvl, e)
+    finally:
+        if old_lang is not None:
+            os.environ['LANG'] = old_lang
+
     msg = try_decode_utf8(res.stdout, 'output from command', logger)
     debug_output(msg)
     return msg.rstrip()