+++ kibot/misc.py
@@ -345,6 +345,7 @@ W_NOVISLA = '(W176) '
 W_IBOMNOCHK = '(W177) '
 W_PREREDEF = '(W178) '
 W_BURIEDVIAS = '(W179) '
+W_BADGITREPO = '(W180) '
 # Somehow arbitrary, the colors are real, but can be different
 PCB_MAT_COLORS = {'fr1': "937042", 'fr2': "949d70", 'fr3': "adacb4", 'fr4': "332B16", 'fr5': "6cc290"}
 PCB_FINISH_COLORS = {'hal': "8b898c", 'hasl': "8b898c", 'imag': "8b898c", 'enig': "cfb96e", 'enepig': "cfb96e",