+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BoM: descriptive error when using `kicad` format and no options found.
 - DRC/ERC: Configurable logo (#889)
 - iBoM: `mark_when_checked` option (#881)
+- KiRi: Support for uninitialized submodules (#888)
 - Report: `buried_vias_count` for KiCad 10
 
 ### Fixed